feat: Bank를 class로 변경, Banker 타입 삭제
