+++ BankManager/Tests/BankManagerTests/BankTests.swift
@@ -27,15 +27,9 @@ final class BankTests: XCTestCase {
         XCTAssertFalse(sut.waitingCustomers.isEmpty)
     }
     
-    func test_은행원추가가_정상적으로이루어지는지_확인할수있다() {
-        sut.addBanker(Banker(id: 1))
-        
-        XCTAssertFalse(sut.bankers.isEmpty)
-    }
-    
     func test_은행업무가_정상적으로이루어지는지_확인할수있다() async throws {
         sut.addCustomer()
-        try await sut.preceedBankWork(Banker(id: 1))
+        try await sut.preceedBankWork()
         
         XCTAssertTrue(sut.waitingCustomers.isEmpty)
     }
